Fix for MTi
The `parse_MTData` method in `mtdevice.py` used for legacy devices perform a test to make sure that all data has been parsed.
The comparison (`!=`) was done between a byte array and an empty string, which always evaluated to true because of the different types.
I suspect this might be from the move from python2 to python3.
Using `len() != 0 ` is a more robust solution.
An other solution would have been to use an empty byte array `b''` for the test.

In the method `fill_from_Sample` of `mtnode.py`, there was an attempt to fill the `seq` field of the `Header` message.
This field has been removed with ROS2 : http://design.ros2.org/articles/changes.html .
As it doesn't seems to be used anywhere else in the project, I just commented the line trying to acces the field.
It might be better to remove the function entirely, but I would probably need to alter `parse_MTData` to do so.